requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.1
django-celery==3.3.1
django-extensions==3.0.5
django-filter==2.3.0
django-jsonfield==1.4.0
djangorestframework==3.11.1
drf-extensions==0.6.0
importlib-metadata==1.7.0
kombu==4.6.11
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
